lab2_3_4/mysite/requirements.txt
asgiref==3.6.0
Django==4.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
waitress==2.1.2
